rename from types to enum
