+++ apps/auth/src/users/enums/index.ts
@@ -1,0 +1,1 @@
+export * from './UserRole.enum';
